Extracted markdown logic to custom hook
