+++ lib/components/Editor/CustomExtensions/Markdown/index.js
@@ -1,7 +1,9 @@
-import React, { useEffect, useLayoutEffect, useRef, useState } from "react";
+import React from "react";
 import classNames from "classnames";
 
-import { htmlToMarkdown, markdownToHtml } from "utils/markdown";
+import { markdownToHtml } from "utils/markdown";
+
+import useMarkdownMode from "./useMarkdownMode";
 
 const MarkdownEditor = ({
   editor,
@@ -13,46 +15,17 @@ const MarkdownEditor = ({
   initialValue,
   ...otherProps
 }) => {
-  const [content, setContent] = useState(() =>
-    htmlToMarkdown(editor?.getHTML() || initialValue)
-  );
-
-  const textareaRef = useRef();
-  const contentRef = useRef(content);
-
-  const BORDER_HEIGHT = 1;
-
-  useEffect(() => {
-    contentRef.current = content;
-  }, [content]);
-
-  useEffect(() => {
-    document.addEventListener("keydown", handleSubmit);
-
-    return () => {
-      editor?.commands.setContent(markdownToHtml(contentRef.current));
-      document.removeEventListener("keydown", handleSubmit);
-    };
-  }, []);
-
-  useLayoutEffect(() => {
-    textareaRef.current.style = style;
-    strategy === "flexible" &&
-      (textareaRef.current.style.height =
-        textareaRef.current.scrollHeight + 2 * BORDER_HEIGHT + "px");
-  }, [content]);
+  const { textareaRef, content, setContent } = useMarkdownMode({
+    editor,
+    style,
+    onSubmit,
+    strategy,
+    initialValue,
+  });
 
   const handleChange = (e) => {
     setContent(e.target.value);
     onChange?.(markdownToHtml(e.target.value));
-  };
-
-  // Submit on Ctrl+Enter and Cmd+Enter
-  const handleSubmit = (e) => {
-    if (e.key === "Enter" && (e.metaKey || e.ctrlKey)) {
-      onSubmit?.(markdownToHtml(contentRef.current));
-      textareaRef.current.blur();
-    }
   };
 
   return (
